:truck: api: move PrivateCertificate from lib to src
