+++ api/tests/tooling/domain-builder/factory/build-private-certificate.js
@@ -1,4 +1,4 @@
-import { PrivateCertificate } from '../../../../lib/domain/models/PrivateCertificate.js';
+import { PrivateCertificate } from '../../../../src/certification/results/domain/models/PrivateCertificate.js';
 import { JuryComment, JuryCommentContexts } from '../../../../src/certification/shared/domain/models/JuryComment.js';
 
 const buildPrivateCertificate = function ({
